Another round of refactoring of the Streams related classes - braking ALL the subclasses down into quasi-namespace and 'Ptr' class. 	?	OpenSSLInputStream 	?	OpenSSLOutputStream 	?	FileOutputStream 	?	FileInputStream 	?	SocketStream 	?	InputStreamFromStdIStream 	?	OutputStreamFromStdOStream 	?	BufferedIbnputStream 	?	BufferedOutputStream 	?	ExternallyOwnedMemoryInputStream 	?	SharedMEmeoryStream 	?	MemoryStream
NOT BACKWARD COMPATIBLE, and probably NOT the last related change (New coming):

In calling code to adapt, you may need to change:

	MemoryStream<X> x;
To
	MemoryStream<X>::Ptr x = MemoryStream<X> {};

And Change:
	return InputStreamFromStdIStream<Memory::Byte> (tmpStrm).ReadAll ();
to:
	return InputStream<Memory::Byte>::Ptr{ InputStreamFromStdIStream<Memory::Byte> (tmpStrm) }.ReadAll ();

Passes regtests on windoze, but still must test unix (before trying next stage)...

+++ Library/Sources/Stroika/Foundation/IO/Network/ConnectionOrientedSocket.cpp
@@ -98,7 +98,7 @@ namespace {
                     FD_SET (fSD_, &input);
                     struct timeval timeout {
                     };
-                    if (::select (fSD_ + 1, &input, NULL, NULL, &timeout) == 1) {
+                    if (::select (static_cast<int> (fSD_) + 1, &input, NULL, NULL, &timeout) == 1) {
                         if (intoStart == nullptr) {
                             return 1; // dont know how much, but doesnt matter, since read allows returning just one byte if thats all thats available
                         }
